requirements.txt
altgraph==0.17.5
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
pystray==0.19.5
pywin32==311
pywin32-ctypes==0.2.3
setuptools==80.9.0
shiboken6==6.10.1
six==1.17.0
waitress==3.0.2
Werkzeug==3.1.3
winshell==0.6
